website/backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
pandas==2.2.3